+++ pkg/aoc/aoc.go
@@ -9,4 +9,4 @@ type Runnable interface {
 	Part2(f *os.File) int
 }
 
-type PartFunc func(f *os.File) int
+type PartFunc func(f *os.File) int64
